move from manager to __init__
